Provide CSharp typename resolution at IType level
